backend/tournament-start-round/requirements.txt
psycopg2-binary>=2.9.9
pusher==3.3.2